change to 64bit integer

+++ include/tracker.h
@@ -8,9 +8,9 @@ enum { PEER_ID_LENGTH = 20 };
 typedef struct TrackerRequest {
   info_hash_t info_hash;
   unsigned char peer_id[PEER_ID_LENGTH];
-  uint16_t port;
-  uint16_t uploaded;
-  uint16_t downloaded;
-  uint16_t left;
+  uint64_t port;
+  uint64_t uploaded;
+  uint64_t downloaded;
+  uint64_t left;
   bool compact;
 } tracker_request_t;
